Initial repository README and assets
The `master` branch stores documentations and assets about the
repository itself and the different branches that provide the specific
template repository data.
This initial `master` commit introduces the `README.md` with basic
content as well as the projects logo and repository hero assets.
it will be extended later on with more project information and usage
instruction for the different template branches.

Resolves GH-1