GitHub "CI" Action Workflow
Using GitHub Actions [1] brings many advantages like a
"close-to-the-source" development pipeline. Having the code and
automated pipelines/workflows in one place is worth a lot.
This also comes along with the perfect and smooth integrations into
GitHub platform and page itself like status reports on PRs and many more
possibilities like the customization and triggering of workflows through
webhooks [2] for almost every event that can occur in a
repository/issue/PR etc.

See the official GitHub Actions documentation [3] for details about
setups, features, the configuration API and many more!

This initial implementation focused on "continuous integration" to the
Node.js scripts with Yarn that are defined in the `package.json` file.

There is official support for SVG badges, but in order to use the same
custom configurations and style of already included badges the ones
provided by shields.io [4] are used.

[1]: https://docs.github.com/en/actions
[2]: https://help.github.com/en/articles/events-that-trigger-workflows#webhook-events
[3]: https://help.github.com/en/categories/automating-your-workflow-with-github-actions
[4]: https://shields.io

GH-28