GitHub issue templates
This commit makes use of GitHub's feature to define multiple issue
templates [1]. The initial template file that has been used when the
feature was introduced [2] and is now used as a fallback/generic
template to redirect to one of the other specific templates.
The UI helps users to open a new issue in projects by prompting them to
choose from multiple issue types.

[1]: https://blog.github.com/2018-01-25-multiple-issue-and-pull-request-templates
[2]: https://blog.github.com/2016-02-17-issue-and-pull-request-templates

GH-32