More fine-grained GitHub Action trigger configuration
The CI GitHub Action Workflow used the `push` setting for the `on` field
in order to trigger the workflow. That was superficial and the workflow
was running for every new commit and PR.

To improve this behavior and prevent unnecessary workflow runs the
`push` field [1] of the `on` field [2] has been configured to only run
for the `main` branch and `v*` tags while the `pull_request` field [3]
is set without any specific configuration so that it runs for all PRs
regardless of the target branch.

[1]: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#onpushpull_requestbranchestags
[2]: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#on
[3]: https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#onpushpull_requestpaths

GH-34