Lowercase naming for GitHub Action workflows
Even though it is possible to use uppercase names (including
whitespaces) for GitHub Action workflows it is best practice for almost
every language to use lowercase names. This prevents problems with
parsing as well as errors due to lower- and uppercase mismatches.

One example is the shields.io [1] SVG badge that is used in the README
of this repository: The actual workflow is only found when the name
matches the exact notation including lower- and uppercase characters:

```diff
- https://img.shields.io/github/workflow/status/svengreb/tmpl/ci.svg
+ https://img.shields.io/github/workflow/status/svengreb/tmpl/CI.svg
```

To mitigate such problems the name has been changed to lowercase only.

[1]: https://shields.io

GH-36