Optimize OS and Go version matrix strategy for CI workflow
Before the CI workflow used a matrix strategy to run the `lint-node`
job, but this was not necessary for this repository. It has been
improved to make the workflow run faster by avoiding unnecessary steps:

- The `lint-node` job has been changed to only run on the currently
  latest stable Node version `14.x` [1] only on Linux because this
  repository is not focused on JavaScript but only runs Node based tools
  to lint other files within this repository.

These changes also help to keep the required GitHub Action run minutes
for the account of this repository as small as possible without wasting
resources for unnecessary tasks.

[1]: https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V14.md#14.12.0

GH-46