Prepare for project/repository publication
To switch the GitHub repository visibility [1] to "public" a few
adjustments had to be made.
This includes the referenced SVG images in documentations, like e.g.
the repository hero, to use the URLs for public repositories instead of
the ones that allow to resolve the files in private repositories.

[1]: https://docs.github.com/en/free-pro-team@latest/github/administering-a-repository/setting-repository-visibility

GH-59